requirements.txt
streamlit
pandas
google-genai
plotly
openpyxl
pywebview
scikit-learn
matplotlib
seaborn
